Initial commit
This data processing pipeline will ingest Excel files from data scraping (not provided here), transform data and save the result dataset to a csv format file as well as fetch into MySQL database.  It was developed on Ubuntu 18.04 LTS with Python 3.6.9 and MySQL database 5.7.29.  There are few tools and Python3 packages are required to run it.